io and training compatible with slar and photonlib

+++ algorithms/opticsensor.py
@@ -132,7 +132,7 @@ class dot_steel_PMT():
         else:
             raise ValueError(f"Unknown angular response model: {self.pmt_ang_eff}")
 
-        self.n_PE = n_photons * self.pmt_eff // 1
+        self.n_PE = (n_photons * self.pmt_eff).int()
         self.max_n = int(torch.max(self.n_PE).item())
         self.hit_mask = self.n_PE > 0
